pyproject.toml
[project]
name = "prl_assets"
version = "0.1.0"
description = "Reusable manipulation objects for Personal Robotics Lab projects"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Personal Robotics Lab" }
]
dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mujoco>=3.0",
    "pillow>=10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prl_assets"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
